@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import logging
 
 import numpy as np
@@ -10,11 +12,13 @@ logger = logging.getLogger(__name__)
 
 class AbsTaskSummarization(AbsTask):
     """
-    Abstract class for re-ranking experiments.
-    Child-classes must implement the following properties:
-    self.corpus = {'dev': Dict[id, str], 'test': Dict[id, str]}         #id => sentence
-    self.queries = {'dev': Dict[id, str], 'test': Dict[id, str]}
-    self.relevant_docs = {'dev': Dict[id, set], 'test': Dict[id, set]}
+    Abstract class for summarization experiments.
+
+    self.load_data() must generate a huggingface dataset with a split matching self.metadata_dict["eval_splits"], and assign it to self.dataset. It must contain the following columns:
+        text: str
+        human_summaries: list[str]
+        machine_summaries: list[str]
+        relevance: list[float] (the score of the machine generated summaries)
     """
 
     def __init__(self, **kwargs):
@@ -22,11 +26,11 @@ class AbsTaskSummarization(AbsTask):
 
     @property
     def min_score(self):
-        return self.description["min_score"]
+        return self.metadata_dict["min_score"]
 
     @property
     def max_score(self):
-        return self.description["max_score"]
+        return self.metadata_dict["max_score"]
 
     def evaluate(self, model, split, **kwargs):
         if not self.data_loaded:
@@ -35,11 +39,15 @@ class AbsTaskSummarization(AbsTask):
         if self.is_crosslingual:
             scores = {}
             for lang in self.dataset:
-                logger.info(f"\nTask: {self.description['name']}, split: {split}, language: {lang}. Running...")
+                logger.info(
+                    f"\nTask: {self.metadata_dict['name']}, split: {split}, language: {lang}. Running..."
+                )
                 data_split = self.dataset[lang][split]
                 scores[lang] = self._evaluate_split(model, data_split, **kwargs)
         else:
-            logger.info(f"\nTask: {self.description['name']}, split: {split}. Running...")
+            logger.info(
+                f"\nTask: {self.metadata_dict['name']}, split: {split}. Running..."
+            )
             data_split = self.dataset[split]
             scores = self._evaluate_split(model, data_split, **kwargs)
 
@@ -47,7 +55,11 @@ class AbsTaskSummarization(AbsTask):
 
     def _evaluate_split(self, model, data_split, **kwargs):
         normalized_scores = list(
-            map(lambda x: (np.array(x) - self.min_score) / (self.max_score - self.min_score), data_split["relevance"])
+            map(
+                lambda x: (np.array(x) - self.min_score)
+                / (self.max_score - self.min_score),
+                data_split["relevance"],
+            )
         )
         evaluator = SummarizationEvaluator(
             machine_summaries=data_split["machine_summaries"],
