@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import logging
 from collections import defaultdict
 
@@ -16,29 +18,55 @@ logger = logging.getLogger(__name__)
 class AbsTaskClassification(AbsTask):
     """
     Abstract class for kNN classification tasks
-    The similarity is computed between pairs and the results are ranked. Average precision
-    is computed to measure how well the methods can be used for classification. #TODO:
+    The similarity is computed between pairs and the results are ranked.
+
+    self.load_data() must generate a huggingface dataset with a split matching self.metadata_dict["eval_splits"], and assign it to self.dataset. It must contain the following columns:
+        text: str
+        label: int
     """
 
-    def __init__(self, method="logReg", n_experiments=None, samples_per_label=None, k=3, batch_size=32, **kwargs):
+    def __init__(
+        self,
+        method="logReg",
+        n_experiments=None,
+        samples_per_label=None,
+        k=3,
+        batch_size=32,
+        **kwargs,
+    ):
         super().__init__(**kwargs)
         self.batch_size = batch_size
         self.method = method
 
         # Bootstrap parameters
-        self.n_experiments = n_experiments if n_experiments is not None else self.description.get("n_experiments", 10)
+        self.n_experiments = (
+            n_experiments
+            if n_experiments is not None
+            else self.metadata_dict.get("n_experiments", 10)
+        )
         self.samples_per_label = (
-            samples_per_label if samples_per_label is not None else self.description.get("samples_per_label", 8)
+            samples_per_label
+            if samples_per_label is not None
+            else self.metadata_dict.get("samples_per_label", 8)
         )
 
         # kNN parameters
         self.k = k
 
+        # Run metadata validation by instantiating addressing the attribute
+        # This is quite hacky. Ideally, this would be done in the constructor of
+        # each concrete task, but then we have to duplicate the __init__ method's
+        # interface.
+        if hasattr(self, "metadata"):
+            self.metadata
+
     def _add_main_score(self, scores):
-        if self.description["main_score"] in scores:
-            scores["main_score"] = scores[self.description["main_score"]]
+        if self.metadata_dict["main_score"] in scores:
+            scores["main_score"] = scores[self.metadata_dict["main_score"]]
         else:
-            logger.warn(f"main score {self.description['main_score']} not found in scores {scores.keys()}")
+            logger.warn(
+                f"main score {self.metadata_dict['main_score']} not found in scores {scores.keys()}"
+            )
 
     def evaluate(self, model, eval_split="test", train_split="train", **kwargs):
         if not self.data_loaded:
@@ -47,26 +75,41 @@ class AbsTaskClassification(AbsTask):
         if self.is_multilingual:
             scores = {}
             for lang in self.dataset:
-                logger.info(f"\nTask: {self.description['name']}, split: {eval_split}, language: {lang}. Running...")
-                scores[lang] = self._evaluate_monolingual(model, self.dataset[lang], eval_split, train_split, **kwargs)
+                logger.info(
+                    f"\nTask: {self.metadata_dict['name']}, split: {eval_split}, language: {lang}. Running..."
+                )
+                scores[lang] = self._evaluate_monolingual(
+                    model, self.dataset[lang], eval_split, train_split, **kwargs
+                )
                 self._add_main_score(scores[lang])
         else:
-            logger.info(f"\nTask: {self.description['name']}, split: {eval_split}. Running...")
-            scores = self._evaluate_monolingual(model, self.dataset, eval_split, train_split, **kwargs)
+            logger.info(
+                f"\nTask: {self.metadata_dict['name']}, split: {eval_split}. Running..."
+            )
+            scores = self._evaluate_monolingual(
+                model, self.dataset, eval_split, train_split, **kwargs
+            )
             self._add_main_score(scores)
 
         return scores
 
-    def _evaluate_monolingual(self, model, dataset, eval_split="test", train_split="train", **kwargs):
+    def _evaluate_monolingual(
+        self, model, dataset, eval_split="test", train_split="train", **kwargs
+    ):
         train_split = dataset[train_split]
         eval_split = dataset[eval_split]
         params = {"k": self.k, "batch_size": self.batch_size}
         params.update(kwargs)
 
         scores = []
-        test_cache, idxs = None, None  # we store idxs to make the shuffling reproducible
+        test_cache, idxs = (
+            None,
+            None,
+        )  # we store idxs to make the shuffling reproducible
         for i in range(self.n_experiments):
-            logger.info("=" * 10 + f" Experiment {i+1}/{self.n_experiments} " + "=" * 10)
+            logger.info(
+                "=" * 10 + f" Experiment {i+1}/{self.n_experiments} " + "=" * 10
+            )
             # Bootstrap `self.samples_per_label` samples per label for each split
             X_sampled, y_sampled, idxs = self._undersample_data(
                 train_split["text"], train_split["label"], self.samples_per_label, idxs
@@ -74,15 +117,27 @@ class AbsTaskClassification(AbsTask):
 
             if self.method == "kNN":
                 evaluator = kNNClassificationEvaluator(
-                    X_sampled, y_sampled, eval_split["text"], eval_split["label"], **params
+                    X_sampled,
+                    y_sampled,
+                    eval_split["text"],
+                    eval_split["label"],
+                    **params,
                 )
             elif self.method == "kNN-pytorch":
                 evaluator = kNNClassificationEvaluatorPytorch(
-                    X_sampled, y_sampled, eval_split["text"], eval_split["label"], **params
+                    X_sampled,
+                    y_sampled,
+                    eval_split["text"],
+                    eval_split["label"],
+                    **params,
                 )
             elif self.method == "logReg":
                 evaluator = logRegClassificationEvaluator(
-                    X_sampled, y_sampled, eval_split["text"], eval_split["label"], **params
+                    X_sampled,
+                    y_sampled,
+                    eval_split["text"],
+                    eval_split["label"],
+                    **params,
                 )
             else:
                 raise ValueError(f"Method {self.method} not supported")
@@ -94,7 +149,9 @@ class AbsTaskClassification(AbsTask):
             return scores[0]
         else:
             avg_scores = {k: np.mean([s[k] for s in scores]) for k in scores[0].keys()}
-            std_errors = {k + "_stderr": np.std([s[k] for s in scores]) for k in scores[0].keys()}
+            std_errors = {
+                k + "_stderr": np.std([s[k] for s in scores]) for k in scores[0].keys()
+            }
             return {**avg_scores, **std_errors}
 
     def _undersample_data(self, X, y, samples_per_label, idxs=None):
