@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import logging
 
 from ..evaluation.evaluators import STSEvaluator
@@ -5,13 +7,15 @@ from .AbsTask import AbsTask
 
 logger = logging.getLogger(__name__)
 
+
 class AbsTaskSTS(AbsTask):
     """
-    Abstract class for re-ranking experiments.
-    Child-classes must implement the following properties:
-    self.corpus = {'dev': Dict[id, str], 'test': Dict[id, str]}         #id => sentence
-    self.queries = {'dev': Dict[id, str], 'test': Dict[id, str]}
-    self.relevant_docs = {'dev': Dict[id, set], 'test': Dict[id, set]}
+    Abstract class for STS experiments.
+
+    self.load_data() must generate a huggingface dataset with a split matching self.metadata_dict["eval_splits"], and assign it to self.dataset. It must contain the following columns::
+        sentence1: str
+        sentence2: str
+        score: float
     """
 
     def __init__(self, **kwargs):
@@ -19,11 +23,11 @@ class AbsTaskSTS(AbsTask):
 
     @property
     def min_score(self):
-        return self.description["min_score"]
+        return self.metadata_dict["min_score"]
 
     @property
     def max_score(self):
-        return self.description["max_score"]
+        return self.metadata_dict["max_score"]
 
     def evaluate(self, model, split, **kwargs):
         if not self.data_loaded:
@@ -32,11 +36,15 @@ class AbsTaskSTS(AbsTask):
         if self.is_crosslingual or self.is_multilingual:
             scores = {}
             for lang in self.dataset:
-                logger.info(f"Task: {self.description['name']}, split: {split}, language: {lang}. Running...")
+                logger.info(
+                    f"Task: {self.metadata_dict['name']}, split: {split}, language: {lang}. Running..."
+                )
                 data_split = self.dataset[lang][split]
                 scores[lang] = self._evaluate_split(model, data_split, **kwargs)
         else:
-            logger.info(f"\nTask: {self.description['name']}, split: {split}. Running...")
+            logger.info(
+                f"\nTask: {self.metadata_dict['name']}, split: {split}. Running..."
+            )
             data_split = self.dataset[split]
             scores = self._evaluate_split(model, data_split, **kwargs)
 
@@ -47,6 +55,11 @@ class AbsTaskSTS(AbsTask):
             return (x - self.min_score) / (self.max_score - self.min_score)
 
         normalized_scores = list(map(normalize, data_split["score"]))
-        evaluator = STSEvaluator(data_split["sentence1"], data_split["sentence2"], normalized_scores, **kwargs)
+        evaluator = STSEvaluator(
+            data_split["sentence1"],
+            data_split["sentence2"],
+            normalized_scores,
+            **kwargs,
+        )
         metrics = evaluator(model)
         return metrics
