@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import logging
 
 import numpy as np
@@ -11,7 +13,9 @@ logger = logging.getLogger(__name__)
 
 
 class BitextMiningEvaluator(Evaluator):
-    def __init__(self, sentences1, sentences2, gold, batch_size=32, limit=None, **kwargs):
+    def __init__(
+        self, sentences1, sentences2, gold, batch_size=32, limit=None, **kwargs
+    ):
         super().__init__(**kwargs)
         self.gold = gold
         self.sentences1 = [sentences1[i] for (i, j) in self.gold]
@@ -91,28 +95,41 @@ class BitextMiningEvaluator(Evaluator):
             for corpus_start_idx in range(0, len(corpus_embeddings), corpus_chunk_size):
                 # Compute cosine similarities
                 cos_scores = score_function(
-                    query_embeddings[query_start_idx : query_start_idx + query_chunk_size],
-                    corpus_embeddings[corpus_start_idx : corpus_start_idx + corpus_chunk_size],
+                    query_embeddings[
+                        query_start_idx : query_start_idx + query_chunk_size
+                    ],
+                    corpus_embeddings[
+                        corpus_start_idx : corpus_start_idx + corpus_chunk_size
+                    ],
                 )
 
                 # Get top-k scores
                 cos_scores_top_k_values, cos_scores_top_k_idx = torch.topk(
-                    cos_scores, min(top_k, len(cos_scores[0])), dim=1, largest=True, sorted=False
+                    cos_scores,
+                    min(top_k, len(cos_scores[0])),
+                    dim=1,
+                    largest=True,
+                    sorted=False,
                 )
                 cos_scores_top_k_values = cos_scores_top_k_values.cpu().tolist()
                 cos_scores_top_k_idx = cos_scores_top_k_idx.cpu().tolist()
 
                 for query_itr in range(len(cos_scores)):
                     for sub_corpus_id, score in zip(
-                        cos_scores_top_k_idx[query_itr], cos_scores_top_k_values[query_itr]
+                        cos_scores_top_k_idx[query_itr],
+                        cos_scores_top_k_values[query_itr],
                     ):
                         corpus_id = corpus_start_idx + sub_corpus_id
                         query_id = query_start_idx + query_itr
-                        queries_result_list[query_id].append({"corpus_id": corpus_id, "score": score})
+                        queries_result_list[query_id].append(
+                            {"corpus_id": corpus_id, "score": score}
+                        )
 
         # Sort and strip to top_k results
         for idx in range(len(queries_result_list)):
-            queries_result_list[idx] = sorted(queries_result_list[idx], key=lambda x: x["score"], reverse=True)
+            queries_result_list[idx] = sorted(
+                queries_result_list[idx], key=lambda x: x["score"], reverse=True
+            )
             queries_result_list[idx] = queries_result_list[idx][0:top_k]
 
         return queries_result_list
