@@ -1,6 +1,9 @@
-from .BornholmskBitextMining import *
-from .BUCCBitextMining import *
-from .DiaBLaBitextMining import *
-from .FloresBitextMining import *
-from .NorwegianCourtsBitextMining import *
-from .TatoebaBitextMining import *
+from __future__ import annotations
+
+from .dan.BornholmskBitextMining import *
+from .multilingual.BUCCBitextMining import *
+from .multilingual.DiaBLaBitextMining import *
+from .multilingual.FloresBitextMining import *
+from .multilingual.norwegian_courts_bitext_mining import *
+from .multilingual.NorwegianCourtsBitextMining import *
+from .multilingual.TatoebaBitextMining import *
