@@ -1,25 +1,29 @@
-from .AmazonCounterfactualClassification import *
-from .AmazonPolarityClassification import *
-from .AmazonReviewsClassification import *
-from .AngryTweetsClassification import *
-from .Banking77Classification import *
-from .CMTEBClassification import *
-from .DalajClassification import *
-from .DanishPoliticalCommentsClassification import *
-from .DKHateClassification import *
-from .EmotionClassification import *
-from .ImdbClassification import *
-from .LccSentimentClassification import *
-from .MasakhaNEWSClassification import *
-from .MassiveIntentClassification import *
-from .MassiveScenarioClassification import *
-from .MTOPDomainClassification import *
-from .MTOPIntentClassification import *
-from .NordicLangClassification import *
-from .NoRecClassification import *
-from .NorwegianParliamentClassification import *
-from .PolishClassification import *
-from .ScalaClassification import *
-from .SweRecClassification import *
-from .ToxicConversationsClassification import *
-from .TweetSentimentExtractionClassification import *
+from __future__ import annotations
+
+from .dan.AngryTweetsClassification import *
+from .dan.DalajClassification import *
+from .dan.DanishPoliticalCommentsClassification import *
+from .dan.DKHateClassification import *
+from .dan.LccSentimentClassification import *
+from .eng.AmazonPolarityClassification import *
+from .eng.Banking77Classification import *
+from .eng.EmotionClassification import *
+from .eng.ImdbClassification import *
+from .eng.NewsClassification import *
+from .eng.ToxicConversationsClassification import *
+from .eng.TweetSentimentExtractionClassification import *
+from .hin.HindiDiscourseClassification import *
+from .multilingual.AmazonCounterfactualClassification import *
+from .multilingual.AmazonReviewsClassification import *
+from .multilingual.MasakhaNEWSClassification import *
+from .multilingual.MassiveIntentClassification import *
+from .multilingual.MassiveScenarioClassification import *
+from .multilingual.MTOPDomainClassification import *
+from .multilingual.MTOPIntentClassification import *
+from .multilingual.NordicLangClassification import *
+from .multilingual.ScalaClassification import *
+from .nob.NoRecClassification import *
+from .nob.NorwegianParliamentClassification import *
+from .pol.PolishClassification import *
+from .swe.SweRecClassification import *
+from .zho.CMTEBClassification import *
