@@ -1,28 +1,33 @@
-from .AlloProfClusteringP2P import *
-from .AlloProfClusteringS2S import *
-from .ArxivClusteringP2P import *
-from .ArxivClusteringS2S import *
-from .BigPatentClustering import *
-from .BiorxivClusteringP2P import *
-from .BiorxivClusteringS2S import *
-from .BlurbsClusteringP2P import *
-from .BlurbsClusteringS2S import *
-from .CMTEBClustering import *
-from .HALClusteringS2S import *
-from .MasakhaNEWSClusteringP2P import *
-from .MasakhaNEWSClusteringS2S import *
-from .MedrxivClusteringP2P import *
-from .MedrxivClusteringS2S import *
-from .MLSUMClusteringP2P import *
-from .MLSUMClusteringS2S import *
-from .PolishClustering import *
-from .RedditClustering import *
-from .RedditClusteringP2P import *
-from .StackExchangeClustering import *
-from .StackExchangeClusteringP2P import *
-from .TenKGnadClusteringP2P import *
-from .TenKGnadClusteringS2S import *
-from .TwentyNewsgroupsClustering import *
-from .WikiCitiesClustering import *
-from .FloresClusteringS2S import *
-from .SpanishNewsClusteringP2P import *
+from __future__ import annotations
+
+from .deu.BlurbsClusteringP2P import *
+from .deu.BlurbsClusteringS2S import *
+from .deu.TenKGnadClusteringP2P import *
+from .deu.TenKGnadClusteringS2S import *
+from .eng.ArxivClusteringP2P import *
+from .eng.ArxivClusteringS2S import *
+from .eng.BigPatentClustering import *
+from .eng.BiorxivClusteringP2P import *
+from .eng.BiorxivClusteringS2S import *
+from .eng.MedrxivClusteringP2P import *
+from .eng.MedrxivClusteringS2S import *
+from .eng.RedditClustering import *
+from .eng.RedditClusteringP2P import *
+from .eng.StackExchangeClustering import *
+from .eng.StackExchangeClusteringP2P import *
+from .eng.TwentyNewsgroupsClustering import *
+from .eng.WikiCitiesClustering import *
+from .fra.AlloProfClusteringP2P import *
+from .fra.AlloProfClusteringS2S import *
+from .fra.HALClusteringS2S import *
+from .fra.MLSUMClusteringP2P import *
+from .fra.MLSUMClusteringS2S import *
+from .multilingual.MasakhaNEWSClusteringP2P import *
+from .multilingual.MasakhaNEWSClusteringS2S import *
+from .nob.snl_clustering import *
+from .nob.vg_clustering import *
+from .pol.PolishClustering import *
+from .spa.FloresClusteringS2S import *
+from .spa.SpanishNewsClusteringP2P import *
+from .swe.swedn_clustering import *
+from .zho.CMTEBClustering import *
