@@ -1,9 +1,10 @@
-from .AlloprofReranking import *
-from .AskUbuntuDupQuestions import *
-from .CMTEBReranking import *
-from .MindSmallReranking import *
-from .CMTEBReranking import *
-from .MIRACLReranking import *
-from .SciDocsReranking import *
-from .StackOverflowDupQuestions import *
-from .SyntecReranking import *
+from __future__ import annotations
+
+from .eng.AskUbuntuDupQuestions import *
+from .eng.MindSmallReranking import *
+from .eng.SciDocsReranking import *
+from .eng.StackOverflowDupQuestions import *
+from .fra.AlloprofReranking import *
+from .fra.SyntecReranking import *
+from .multilingual.MIRACLReranking import *
+from .zho.CMTEBReranking import *
