@@ -1,60 +1,75 @@
-from .AlloprofRetrieval import *
-from .ArguAnaPLRetrieval import *
-from .ArguAnaRetrieval import *
-from .BSARDRetrieval import *
-from .ClimateFEVERRetrieval import *
-from .CMTEBRetrieval import *
-from .CQADupstackAndroidRetrieval import *
-from .CQADupstackEnglishRetrieval import *
-from .CQADupstackGamingRetrieval import *
-from .CQADupstackGisRetrieval import *
-from .CQADupstackMathematicaRetrieval import *
-from .CQADupstackPhysicsRetrieval import *
-from .CQADupstackProgrammersRetrieval import *
-from .CQADupstackStatsRetrieval import *
-from .CQADupstackTexRetrieval import *
-from .CQADupstackUnixRetrieval import *
-from .CQADupstackWebmastersRetrieval import *
-from .CQADupstackWordpressRetrieval import *
-from .DBPediaPLRetrieval import *
-from .DBPediaRetrieval import *
-from .FEVERRetrieval import *
-from .FiQA2018Retrieval import *
-from .FiQAPLRetrieval import *
-from .GermanQuADRetrieval import *
-from .HagridRetrieval import *
-from .HotpotQAPLRetrieval import *
-from .HotpotQARetrieval import *
-from .MSMARCOPLRetrieval import *
-from .MSMARCORetrieval import *
-from .MSMARCOv2Retrieval import *
-from .NarrativeQARetrieval import *
-from .NFCorpusPLRetrieval import *
-from .NFCorpusRetrieval import *
-from .NQPLRetrieval import *
-from .NQRetrieval import *
-from .QuoraPLRetrieval import *
-from .QuoraRetrieval import *
-from .SCIDOCSPLRetrieval import *
-from .SCIDOCSRetrieval import *
-from .SciFactPLRetrieval import *
-from .SciFactRetrieval import *
-from .SyntecRetrieval import *
-from .Touche2020Retrieval import *
-from .TRECCOVIDPLRetrieval import *
-from .TRECCOVIDRetrieval import *
-from .NarrativeQARetrieval import *
-from .GermanQuADRetrieval import *
-from .GerDaLIRRetrieval import *
-from .GermanDPRRetrieval import *
-from .GerDaLIRRetrieval import *
-from .MultiLongDocRetrieval import *
-from .KoStrategyQA import *
-from .KoMrtydi import *
-from .KoMiracl import *
-from .SpanishPassageRetrievalS2S import *
-from .SpanishPassageRetrievalS2P import *
-from .XPQARetrieval import *
-from .MintakaRetrieval import *
-from .MIRACLRetrieval import *
-from .XMarketRetrieval import *
+from __future__ import annotations
+
+from .dan.dan_fever import *
+from .dan.t2nord_retrieval import *
+from .dan.twitterhjerne import *
+from .deu.GerDaLIRRetrieval import *
+from .deu.GerDaLIRSmallRetrieval import *
+from .deu.GermanDPRRetrieval import *
+from .deu.GermanQuADRetrieval import *
+from .deu.LegalQuADRetrieval import *
+from .eng.AILACasedocsRetrieval import *
+from .eng.AILAStatutesRetrieval import *
+from .eng.ArguAnaRetrieval import *
+from .eng.ClimateFEVERRetrieval import *
+from .eng.CQADupstackAndroidRetrieval import *
+from .eng.CQADupstackEnglishRetrieval import *
+from .eng.CQADupstackGamingRetrieval import *
+from .eng.CQADupstackGisRetrieval import *
+from .eng.CQADupstackMathematicaRetrieval import *
+from .eng.CQADupstackPhysicsRetrieval import *
+from .eng.CQADupstackProgrammersRetrieval import *
+from .eng.CQADupstackStatsRetrieval import *
+from .eng.CQADupstackTexRetrieval import *
+from .eng.CQADupstackUnixRetrieval import *
+from .eng.CQADupstackWebmastersRetrieval import *
+from .eng.CQADupstackWordpressRetrieval import *
+from .eng.DBPediaRetrieval import *
+from .eng.FEVERRetrieval import *
+from .eng.FiQA2018Retrieval import *
+from .eng.HagridRetrieval import *
+from .eng.HotpotQARetrieval import *
+from .eng.LegalBenchConsumerContractsQARetrieval import *
+from .eng.LegalBenchCorporateLobbyingRetrieval import *
+from .eng.LegalSummarizationRetrieval import *
+from .eng.MedicalQARetrieval import *
+from .eng.MSMARCORetrieval import *
+from .eng.MSMARCOv2Retrieval import *
+from .eng.NarrativeQARetrieval import *
+from .eng.NFCorpusRetrieval import *
+from .eng.NQRetrieval import *
+from .eng.QuoraRetrieval import *
+from .eng.SCIDOCSRetrieval import *
+from .eng.SciFactRetrieval import *
+from .eng.Touche2020Retrieval import *
+from .eng.TRECCOVIDRetrieval import *
+from .fra.AlloprofRetrieval import *
+from .fra.BSARDRetrieval import *
+from .fra.SyntecRetrieval import *
+from .kor.KoMiracl import *
+from .kor.KoMrtydi import *
+from .kor.KoStrategyQA import *
+from .multilingual.MintakaRetrieval import *
+from .multilingual.MIRACLRetrieval import *
+from .multilingual.MultiLongDocRetrieval import *
+from .multilingual.XMarketRetrieval import *
+from .multilingual.XPQARetrieval import *
+from .nob.norquad import *
+from .nob.snl_retrieval import *
+from .pol.ArguAnaPLRetrieval import *
+from .pol.DBPediaPLRetrieval import *
+from .pol.FiQAPLRetrieval import *
+from .pol.HotpotQAPLRetrieval import *
+from .pol.MSMARCOPLRetrieval import *
+from .pol.NFCorpusPLRetrieval import *
+from .pol.NQPLRetrieval import *
+from .pol.QuoraPLRetrieval import *
+from .pol.SCIDOCSPLRetrieval import *
+from .pol.SciFactPLRetrieval import *
+from .pol.TRECCOVIDPLRetrieval import *
+from .spa.SpanishPassageRetrievalS2P import *
+from .spa.SpanishPassageRetrievalS2S import *
+from .swe.swedn_retrieval import *
+from .swe.swefaq_retrieval import *
+from .zho.CMTEBRetrieval import *
+from .zho.LeCaRDv2Retrieval import *
