@@ -1,2 +1,4 @@
-from .SummEvalFrSummarization import *
-from .SummEvalSummarization import *
+from __future__ import annotations
+
+from .eng.SummEvalSummarization import *
+from .fra.SummEvalFrSummarization import *
