@@ -1,12 +1,12 @@
+from __future__ import annotations
+
 import gzip
 import os
 
 import datasets
+import jsonlines
 import numpy as np
 from tqdm import tqdm
-
-import jsonlines
-
 
 np.random.seed(28042000)
 
