@@ -2,13 +2,13 @@
 Fetch data from https://api.biorxiv.org/ and keep useful information
 """
 
-import gzip
+from __future__ import annotations
 
-from tqdm import tqdm
+import gzip
 
 import jsonlines
 import requests
-
+from tqdm import tqdm
 
 api = "https://api.biorxiv.org/details/biorxiv/2021-01-01/2022-05-10/"
 
