@@ -1,9 +1,9 @@
+from __future__ import annotations
+
 import os
 
 from datasets import load_dataset
-
 from huggingface_hub import create_repo, upload_file
-
 
 repo_name = "imdb"
 create_repo(repo_name, organization="mteb", repo_type="dataset")
@@ -14,5 +14,10 @@ raw_dset = raw_dset.map(lambda x: {"label_text": id2label[x["label"]]}, num_proc
 for split, dset in raw_dset.items():
     save_path = f"{split}.jsonl"
     dset.to_json(save_path)
-    upload_file(path_or_fileobj=save_path, path_in_repo=save_path, repo_id="mteb/" + repo_name, repo_type="dataset")
+    upload_file(
+        path_or_fileobj=save_path,
+        path_in_repo=save_path,
+        repo_id="mteb/" + repo_name,
+        repo_type="dataset",
+    )
     os.system(f"rm {save_path}")
