@@ -1,9 +1,12 @@
 """Example script for benchmarking all datasets constituting the MTEB English leaderboard & average scores"""
 
+from __future__ import annotations
+
 import logging
 
-from mteb import MTEB
 from sentence_transformers import SentenceTransformer
+
+from mteb import MTEB
 
 logging.basicConfig(level=logging.INFO)
 
@@ -109,5 +112,9 @@ model = SentenceTransformer(model_name)
 for task in TASK_LIST:
     logger.info(f"Running task: {task}")
     eval_splits = ["dev"] if task == "MSMARCO" else ["test"]
-    evaluation = MTEB(tasks=[task], task_langs=["en"])  # Remove "en" for running all languages
-    evaluation.run(model, output_folder=f"results/{model_name}", eval_splits=eval_splits)
+    evaluation = MTEB(
+        tasks=[task], task_langs=["en"]
+    )  # Remove "en" for running all languages
+    evaluation.run(
+        model, output_folder=f"results/{model_name}", eval_splits=eval_splits
+    )
