@@ -1,5 +1,7 @@
 """Example script for benchmarking all datasets constituting the MTEB French leaderboard & average scores"""
 
+from __future__ import annotations
+
 import logging
 
 from sentence_transformers import SentenceTransformer
@@ -34,33 +36,17 @@ TASK_LIST_PAIR_CLASSIFICATION = [
     "PawsX",
 ]
 
-TASK_LIST_RERANKING = [
-    "SyntecReranking",
-    "AlloprofReranking"
-]
+TASK_LIST_RERANKING = ["SyntecReranking", "AlloprofReranking"]
 
 TASK_LIST_RETRIEVAL = [
-    "AlloprofRetrieval", 
-    "BSARDRetrieval", 
+    "AlloprofRetrieval",
+    "BSARDRetrieval",
     "SyntecRetrieval",
     "XPQARetrieval",
     "MintakaRetrieval",
 ]
 
-TASK_LIST_STS = [
-    "SummEvalFr",
-    "STSBenchmarkMultilingualSTS",
-    "STS22",
-    "SICKFr"
-]
-
-TASK_LIST_BITEXTMINING = [
-    "DiaBLaBitextMining",
-    "FloresBitextMining",
-    "TatoebaBitextMining",
-    "BUCCBitextMining"
-]
-
+TASK_LIST_STS = ["SummEvalFr", "STSBenchmarkMultilingualSTS", "STS22", "SICKFr"]
 
 TASK_LIST = (
     TASK_LIST_CLASSIFICATION
@@ -69,7 +55,6 @@ TASK_LIST = (
     + TASK_LIST_RERANKING
     + TASK_LIST_RETRIEVAL
     + TASK_LIST_STS
-    + TASK_LIST_BITEXTMINING
 )
 
 model_name = "dangvantuan/sentence-camembert-base"
@@ -78,5 +63,7 @@ model = SentenceTransformer(model_name)
 logger.info(f"Task list : {TASK_LIST}")
 for task in TASK_LIST:
     logger.info(f"Running task: {task}")
-    evaluation = MTEB(tasks=[task], task_langs=["fr"])  # Remove "fr" for running all languages
+    evaluation = MTEB(
+        tasks=[task], task_langs=["fr"]
+    )  # Remove "fr" for running all languages
     evaluation.run(model, output_folder=f"results/{model_name}")
