@@ -1,9 +1,12 @@
 """Example script for benchmarking all datasets constituting the MTEB Polish leaderboard & average scores"""
 
+from __future__ import annotations
+
 import logging
 
-from mteb import MTEB
 from sentence_transformers import SentenceTransformer
+
+from mteb import MTEB
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger("main")
@@ -15,30 +18,16 @@ classification_tasks = [
     "AllegroReviews",
     "PAC",
     "MassiveIntentClassification",
-    "MassiveScenarioClassification"
+    "MassiveScenarioClassification",
 ]
 
-clustering_tasks = [
-    "8TagsClustering"
-]
+clustering_tasks = ["8TagsClustering"]
 
-pair_classification_tasks = [
-    "SICK-E-PL",
-    "PPC",
-    "CDSC-E",
-    "PSC"
-]
+pair_classification_tasks = ["SICK-E-PL", "PPC", "CDSC-E", "PSC"]
 
-sts_tasks = [
-    "SICK-R-PL",
-    "CDSC-R",
-    "STS22"
-]
+sts_tasks = ["SICK-R-PL", "CDSC-R", "STS22"]
 
-tasks = classification_tasks \
-        + clustering_tasks \
-        + pair_classification_tasks \
-        + sts_tasks
+tasks = classification_tasks + clustering_tasks + pair_classification_tasks + sts_tasks
 
 model_name = "sdadas/st-polish-paraphrase-from-distilroberta"
 model = SentenceTransformer(model_name)
